requirements.txt
requests>=2.32.0
httpx>=0.28.1
uvicorn
pylint
pytest
fastapi[standard]>=0.115.9
python-dotenv
prometheus-fastapi-instrumentator
redis
pytest-asyncio
pytest-mock
